046 Split services into different files
Single file for the struct, and another for the implementation.
